force casting

+++ src/main/java/org/maxgamer/quickshop/shop/ContainerShop.java
@@ -560,7 +560,7 @@ public class ContainerShop implements Shop {
                     lines[1] = MsgUtil.getMessageOfflinePlayer("signs.stack-selling", player, "" + MsgUtil.getMessageOfflinePlayer("signs.unlimited", player));
                 } else {
                     lines[1] =
-                            MsgUtil.getMessageOfflinePlayer("signs.stack-selling", player, "" + ((this.getRemainingStock() != 0) ? CalculateUtil.divide(this.getRemainingStock(), this.getItem().getAmount()) : 0));
+                            MsgUtil.getMessageOfflinePlayer("signs.stack-selling", player, "" + ((this.getRemainingStock() != 0) ? (int)CalculateUtil.divide(this.getRemainingStock(), this.getItem().getAmount()) : 0));
                 }
             } else {
                 if (this.getRemainingStock() == -1) {
@@ -578,7 +578,7 @@ public class ContainerShop implements Shop {
                     lines[1] = MsgUtil.getMessageOfflinePlayer("signs.stack-buying", player, "" + MsgUtil.getMessageOfflinePlayer("signs.unlimited", player));
                 } else {
                     lines[1] =
-                            MsgUtil.getMessageOfflinePlayer("signs.stack-buying", player, "" + ((this.getRemainingSpace() != 0) ? CalculateUtil.divide(this.getRemainingSpace(), this.getItem().getAmount()) : 0));
+                            MsgUtil.getMessageOfflinePlayer("signs.stack-buying", player, "" + ((this.getRemainingSpace() != 0) ? (int)CalculateUtil.divide(this.getRemainingSpace(), this.getItem().getAmount()) : 0));
                 }
             } else {
                 if (this.getRemainingSpace() == -1) {
